feat(vis): Publish a better version :bulb:
I've publish a better version for this visualization putting more
information about the collection of data, and more conclusions. The
graph now have more edges, that were filter by genre to 50 edges at max,
not filter as whole. I've also made some modifications to the CSS and JS
to show genre of the artists on hover.